fix(scripts): update nav dll in external script
